+++ src/main/kotlin/com/wutsi/codegen/Context.kt
@@ -8,8 +8,17 @@ class Context(
     val basePackage: String,
     val outputDirectory: String = ".${File.separator}out",
     val jdkVersion: String = "11",
-    val githubUser: String? = null
+    val githubUser: String? = null,
+    val herokuApp: String? = null
 ) {
+    companion object {
+        const val SERVICE_LOGGING = "service:logging"
+        const val SERVICE_CACHE = "service:cache"
+        const val SERVICE_DATABASE = "service:database"
+        const val SERVICE_QUEUE = "service:queue"
+    }
+
+    private val services: MutableList<String> = mutableListOf()
     private val typeRegistry = mutableMapOf<String, Type>()
 
     fun register(ref: String, type: Type) {
@@ -17,4 +26,10 @@ class Context(
     }
 
     fun getType(ref: String): Type? = typeRegistry[ref]
+
+    fun addService(service: String) {
+        services.add(service)
+    }
+
+    fun hasService(service: String): Boolean = services.contains(service)
 }